Created a seperate fil for the render stars function and added star rating for each product
